requirements.txt
pyannote.audio
python-dotenv
openai
fastapi
pydantic
psycopg2-binary
supabase
python-jose[cryptography]
passlib[bcrypt]